third_party/WebKit/LayoutTests/fast/layout/scroll-anchoring/anchor-inside-iframe.html

<!DOCTYPE html>
<script src="../../../resources/run-after-layout-and-paint.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<iframe width="700" height="500" srcdoc="
  <!DOCTYPE html>
  <style> body { height: 1000px } div { height: 100px } </style>
  <div id='block1'>abc</div>
  <div id='block2'>def</div>
"></iframe>
<script>
  async_test((t) => {
    var iframeWindow = document.querySelector("iframe").contentWindow;
    iframeWindow.addEventListener("load", () => {
      var block1 = iframeWindow.document.querySelector("#block1");
      iframeWindow.scrollTo(0, 150);

      runAfterLayoutAndPaint(() => {
        block1.style.height = "200px";
        assert_equals(iframeWindow.scrollY, 250);
        t.done();
      });
    });
  }, "Scroll anchoring in an iframe.");
</script>
